bug fixed

+++ api/model/model.go
@@ -15,32 +15,27 @@ type Criticals struct {
 	Author string  `json:"author"`
 }
 
-
 type Error struct {
-	DateTime string		`json:"dateTime"`
-	ParamName string	`json:"paramName"`
-	ParamValue float32	`json:"paramValue"`
-	Message string		`json:"message"`
+	DateTime   string  `json:"dateTime"`
+	ParamName  string  `json:"paramName"`
+	ParamValue float64 `json:"paramValue"`
+	Message    string  `json:"message"`
 }
 
 type Time struct {
-	Hour int
+	Hour   int
 	Minute int
 	Second int
 }
 
-
-
-
-func (time *Time)NextHour(){
-	time.Hour+=1
+func (time *Time) NextHour() {
+	time.Hour += 1
 }
 
-func (time *Time)ToStringHour()string{
-	if time.Hour>9{
-		return strconv.Itoa(time.Hour)+":00:00"
-	}else {
-		return "0"+strconv.Itoa(time.Hour)+":00:00"
+func (time *Time) ToStringHour() string {
+	if time.Hour > 9 {
+		return strconv.Itoa(time.Hour) + ":00:00"
+	} else {
+		return "0" + strconv.Itoa(time.Hour) + ":00:00"
 	}
 }
-
